refactor(api): extract system routes to src/api/system_routes.py
Move 5 system routes (health, correlation, shutdown, skills, api-info)
and the _terminate_current_process helper from api_server.py into a
self-contained route module following the PR #375 pattern.

- system_routes.py defines HealthResponse locally (no shared modules)
- register_system_routes() resolves _security and
  _require_shutdown_authorization via sys.modules fallback
- Re-exports _terminate_current_process for test monkeypatch compat
- api_server.py: -92 lines (3552 → 3471)

AI-Model: idealab/Qwen3.7-Max-DogFooding
AI-Contributed/Feature: 270/270
AI-Contributed/UT: 0/0
